+++ app/src/main/java/com/jondh/bartenderhelp/MainActivity.kt
@@ -8,7 +8,10 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.jondh.bartenderhelp.model.getDrinksFromResources
 import com.jondh.bartenderhelp.ui.theme.BartenderHelpTheme
+import com.jondh.bartenderhelp.ui.view.DrinkCardScreen
+import com.jondh.bartenderhelp.ui.vm.DrinkCardViewModel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
